Спринт-_-1/pereval_project/requirements.txt
fastapi  
uvicorn  
asyncpg                   # Используем для асинхронного взаимодействия с PostgreSQL  
pydantic  
python-dotenv  
pytest                   # Для трейтинга  
httpx                    # Для тестирования HTTP-запросов